refactor(season): Refactor season entity slice on client
- Refactored season entity slice public api
