+++ packages/client/src/entities/season/api/useSeasonServer.ts
@@ -1,11 +1,11 @@
 import { axios } from '~shared/api/axios';
 
-import type { Season } from '~entities/season/type';
-import { URLS } from '~entities/season/api/urls';
+import type { SeasonType } from '~entities/season';
+import { SEASON_URLS } from '~entities/season/api';
 
 export async function useSeasonServer(year: number) {
     try {
-        const { data: season } = await axios.get<Season>(URLS.year(year));
+        const { data: season } = await axios.get<SeasonType>(SEASON_URLS.year(year));
 
         return { season };
     } catch (error: unknown) {
