+++ packages/client/src/entities/season/ui/index.ts
@@ -1,0 +1,1 @@
+export { SeasonListCard } from './list-card';
